Add files via upload
The "button.py" program runs upon the RaspberryPi's boot-up,  functioning as a switch for our color-detection code in our RaspberryPi system. Moreover, "color.py" contains the main algorithm for our system's color-detection. This includes the color-masking for the pillars' colors and the serial communication system for the Arduino and RaspberryPi.